@@ -21,167 +21,80 @@ use Longman\TelegramBot\Entities\Message;
 use Longman\TelegramBot\Entities\ServerResponse;
 use Longman\TelegramBot\Exception\InvalidBotTokenException;
 use Longman\TelegramBot\Exception\TelegramException;
+use Throwable;
 
 /**
  * Class Request
  *
- * @method static ServerResponse getUpdates(array $data)                      Use this method to receive incoming updates using long polling (wiki). An Array
- *     of Update objects is returned.
- * @method static ServerResponse setWebhook(array $data)                      Use this method to specify a url and receive incoming updates via an outgoing
- *     webhook. Whenever there is an update for the bot, we will send an HTTPS POST request to the specified url, containing a JSON-serialized Update. In case
- *     of an unsuccessful request, we will give up after a reasonable amount of attempts. Returns true.
- * @method static ServerResponse deleteWebhook()                              Use this method to remove webhook integration if you decide to switch back to
- *     getUpdates. Returns True on success. Requires no parameters.
- * @method static ServerResponse getWebhookInfo()                             Use this method to get current webhook status. Requires no parameters. On
- *     success, returns a WebhookInfo object. If the bot is using getUpdates, will return an object with the url field empty.
- * @method static ServerResponse getMe()                                      A simple method for testing your bot's auth token. Requires no parameters.
- *     Returns basic information about the bot in form of a User object.
- * @method static ServerResponse forwardMessage(array $data)                  Use this method to forward messages of any kind. On success, the sent Message is
- *     returned.
+ * @method static ServerResponse getUpdates(array $data)                      Use this method to receive incoming updates using long polling (wiki). An Array of Update objects is returned.
+ * @method static ServerResponse setWebhook(array $data)                      Use this method to specify a url and receive incoming updates via an outgoing webhook. Whenever there is an update for the bot, we will send an HTTPS POST request to the specified url, containing a JSON-serialized Update. In case of an unsuccessful request, we will give up after a reasonable amount of attempts. Returns true.
+ * @method static ServerResponse deleteWebhook()                              Use this method to remove webhook integration if you decide to switch back to getUpdates. Returns True on success. Requires no parameters.
+ * @method static ServerResponse getWebhookInfo()                             Use this method to get current webhook status. Requires no parameters. On success, returns a WebhookInfo object. If the bot is using getUpdates, will return an object with the url field empty.
+ * @method static ServerResponse getMe()                                      A simple method for testing your bot's auth token. Requires no parameters. Returns basic information about the bot in form of a User object.
+ * @method static ServerResponse forwardMessage(array $data)                  Use this method to forward messages of any kind. On success, the sent Message is returned.
  * @method static ServerResponse sendPhoto(array $data)                       Use this method to send photos. On success, the sent Message is returned.
- * @method static ServerResponse sendAudio(array $data)                       Use this method to send audio files, if you want Telegram clients to display them
- *     in the music player. Your audio must be in the .mp3 format. On success, the sent Message is returned. Bots can currently send audio files of up to 50 MB
- *     in size, this limit may be changed in the future.
- * @method static ServerResponse sendDocument(array $data)                    Use this method to send general files. On success, the sent Message is returned.
- *     Bots can currently send files of any type of up to 50 MB in size, this limit may be changed in the future.
+ * @method static ServerResponse sendAudio(array $data)                       Use this method to send audio files, if you want Telegram clients to display them in the music player. Your audio must be in the .mp3 format. On success, the sent Message is returned. Bots can currently send audio files of up to 50 MB in size, this limit may be changed in the future.
+ * @method static ServerResponse sendDocument(array $data)                    Use this method to send general files. On success, the sent Message is returned. Bots can currently send files of any type of up to 50 MB in size, this limit may be changed in the future.
  * @method static ServerResponse sendSticker(array $data)                     Use this method to send .webp stickers. On success, the sent Message is returned.
- * @method static ServerResponse sendVideo(array $data)                       Use this method to send video files, Telegram clients support mp4 videos (other
- *     formats may be sent as Document). On success, the sent Message is returned. Bots can currently send video files of up to 50 MB in size, this limit may
- *     be changed in the future.
- * @method static ServerResponse sendAnimation(array $data)                   Use this method to send animation files (GIF or H.264/MPEG-4 AVC video without
- *     sound). On success, the sent Message is returned. Bots can currently send animation files of up to 50 MB in size, this limit may be changed in the
- *     future.
- * @method static ServerResponse sendVoice(array $data)                       Use this method to send audio files, if you want Telegram clients to display the
- *     file as a playable voice message. For this to work, your audio must be in an .ogg file encoded with OPUS (other formats may be sent as Audio or
- *     Document). On success, the sent Message is returned. Bots can currently send voice messages of up to 50 MB in size, this limit may be changed in the
- *     future.
+ * @method static ServerResponse sendVideo(array $data)                       Use this method to send video files, Telegram clients support mp4 videos (other formats may be sent as Document). On success, the sent Message is returned. Bots can currently send video files of up to 50 MB in size, this limit may be changed in the future.
+ * @method static ServerResponse sendAnimation(array $data)                   Use this method to send animation files (GIF or H.264/MPEG-4 AVC video without sound). On success, the sent Message is returned. Bots can currently send animation files of up to 50 MB in size, this limit may be changed in the future.
+ * @method static ServerResponse sendVoice(array $data)                       Use this method to send audio files, if you want Telegram clients to display the file as a playable voice message. For this to work, your audio must be in an .ogg file encoded with OPUS (other formats may be sent as Audio or Document). On success, the sent Message is returned. Bots can currently send voice messages of up to 50 MB in size, this limit may be changed in the future.
  * @method static ServerResponse sendVideoNote(array $data)                   Use this method to send video messages. On success, the sent Message is returned.
- * @method static ServerResponse sendMediaGroup(array $data)                  Use this method to send a group of photos or videos as an album. On success, an
- *     array of the sent Messages is returned.
- * @method static ServerResponse sendLocation(array $data)                    Use this method to send point on the map. On success, the sent Message is
- *     returned.
- * @method static ServerResponse editMessageLiveLocation(array $data)         Use this method to edit live location messages sent by the bot or via the bot
- *     (for inline bots). A location can be edited until its live_period expires or editing is explicitly disabled by a call to stopMessageLiveLocation. On
- *     success, if the edited message was sent by the bot, the edited Message is returned, otherwise True is returned.
- * @method static ServerResponse stopMessageLiveLocation(array $data)         Use this method to stop updating a live location message sent by the bot or via
- *     the bot (for inline bots) before live_period expires. On success, if the message was sent by the bot, the sent Message is returned, otherwise True is
- *     returned.
- * @method static ServerResponse sendVenue(array $data)                       Use this method to send information about a venue. On success, the sent Message
- *     is returned.
+ * @method static ServerResponse sendMediaGroup(array $data)                  Use this method to send a group of photos or videos as an album. On success, an array of the sent Messages is returned.
+ * @method static ServerResponse sendLocation(array $data)                    Use this method to send point on the map. On success, the sent Message is returned.
+ * @method static ServerResponse editMessageLiveLocation(array $data)         Use this method to edit live location messages sent by the bot or via the bot (for inline bots). A location can be edited until its live_period expires or editing is explicitly disabled by a call to stopMessageLiveLocation. On success, if the edited message was sent by the bot, the edited Message is returned, otherwise True is returned.
+ * @method static ServerResponse stopMessageLiveLocation(array $data)         Use this method to stop updating a live location message sent by the bot or via the bot (for inline bots) before live_period expires. On success, if the message was sent by the bot, the sent Message is returned, otherwise True is returned.
+ * @method static ServerResponse sendVenue(array $data)                       Use this method to send information about a venue. On success, the sent Message is returned.
  * @method static ServerResponse sendContact(array $data)                     Use this method to send phone contacts. On success, the sent Message is returned.
- * @method static ServerResponse sendPoll(array $data)                        Use this method to send a native poll. A native poll can't be sent to a private
- *     chat. On success, the sent Message is returned.
- * @method static ServerResponse sendDice(array $data)                        Use this method to send a dice, which will have a random value from 1 to 6. On
- *     success, the sent Message is returned.
- * @method static ServerResponse sendChatAction(array $data)                  Use this method when you need to tell the user that something is happening on the
- *     bot's side. The status is set for 5 seconds or less (when a message arrives from your bot, Telegram clients clear its typing status). Returns True on
- *     success.
- * @method static ServerResponse getUserProfilePhotos(array $data)            Use this method to get a list of profile pictures for a user. Returns a
- *     UserProfilePhotos object.
- * @method static ServerResponse getFile(array $data)                         Use this method to get basic info about a file and prepare it for downloading.
- *     For the moment, bots can download files of up to 20MB in size. On success, a File object is returned. The file can then be downloaded via the link
- *     https://api.telegram.org/file/bot<token>/<file_path>, where <file_path> is taken from the response. It is guaranteed that the link will be valid for at
- *     least 1 hour. When the link expires, a new one can be requested by calling getFile again.
- * @method static ServerResponse kickChatMember(array $data)                  Use this method to kick a user from a group, a supergroup or a channel. In the
- *     case of supergroups and channels, the user will not be able to return to the group on their own using invite links, etc., unless unbanned first. The bot
- *     must be an administrator in the chat for this to work and must have the appropriate admin rights. Returns True on success.
- * @method static ServerResponse unbanChatMember(array $data)                 Use this method to unban a previously kicked user in a supergroup or channel. The
- *     user will not return to the group or channel automatically, but will be able to join via link, etc. The bot must be an administrator for this to work.
- *     Returns True on success.
- * @method static ServerResponse restrictChatMember(array $data)              Use this method to restrict a user in a supergroup. The bot must be an
- *     administrator in the supergroup for this to work and must have the appropriate admin rights. Pass True for all permissions to lift restrictions from a
- *     user. Returns True on success.
- * @method static ServerResponse promoteChatMember(array $data)               Use this method to promote or demote a user in a supergroup or a channel. The bot
- *     must be an administrator in the chat for this to work and must have the appropriate admin rights. Pass False for all boolean parameters to demote a
- *     user. Returns True on success.
- * @method static ServerResponse setChatAdministratorCustomTitle(array $data) Use this method to set a custom title for an administrator in a supergroup
- *     promoted by the bot. Returns True on success.
- * @method static ServerResponse setChatPermissions(array $data)              Use this method to set default chat permissions for all members. The bot must be
- *     an administrator in the group or a supergroup for this to work and must have the can_restrict_members admin rights. Returns True on success.
- * @method static ServerResponse exportChatInviteLink(array $data)            Use this method to generate a new invite link for a chat. Any previously
- *     generated link is revoked. The bot must be an administrator in the chat for this to work and must have the appropriate admin rights. Returns the new
- *     invite link as String on success.
- * @method static ServerResponse setChatPhoto(array $data)                    Use this method to set a new profile photo for the chat. Photos can't be changed
- *     for private chats. The bot must be an administrator in the chat for this to work and must have the appropriate admin rights. Returns True on success.
- * @method static ServerResponse deleteChatPhoto(array $data)                 Use this method to delete a chat photo. Photos can't be changed for private
- *     chats. The bot must be an administrator in the chat for this to work and must have the appropriate admin rights. Returns True on success.
- * @method static ServerResponse setChatTitle(array $data)                    Use this method to change the title of a chat. Titles can't be changed for
- *     private chats. The bot must be an administrator in the chat for this to work and must have the appropriate admin rights. Returns True on success.
- * @method static ServerResponse setChatDescription(array $data)              Use this method to change the description of a group, a supergroup or a channel.
- *     The bot must be an administrator in the chat for this to work and must have the appropriate admin rights. Returns True on success.
- * @method static ServerResponse pinChatMessage(array $data)                  Use this method to pin a message in a supergroup or a channel. The bot must be an
- *     administrator in the chat for this to work and must have the ‘can_pin_messages’ admin right in the supergroup or ‘can_edit_messages’ admin right in the
- *     channel. Returns True on success.
- * @method static ServerResponse unpinChatMessage(array $data)                Use this method to unpin a message in a supergroup or a channel. The bot must be
- *     an administrator in the chat for this to work and must have the ‘can_pin_messages’ admin right in the supergroup or ‘can_edit_messages’ admin right in
- *     the channel. Returns True on success.
- * @method static ServerResponse leaveChat(array $data)                       Use this method for your bot to leave a group, supergroup or channel. Returns
- *     True on success.
- * @method static ServerResponse getChat(array $data)                         Use this method to get up to date information about the chat (current name of the
- *     user for one-on-one conversations, current username of a user, group or channel, etc.). Returns a Chat object on success.
- * @method static ServerResponse getChatAdministrators(array $data)           Use this method to get a list of administrators in a chat. On success, returns an
- *     Array of ChatMember objects that contains information about all chat administrators except other bots. If the chat is a group or a supergroup and no
- *     administrators were appointed, only the creator will be returned.
+ * @method static ServerResponse sendPoll(array $data)                        Use this method to send a native poll. A native poll can't be sent to a private chat. On success, the sent Message is returned.
+ * @method static ServerResponse sendDice(array $data)                        Use this method to send a dice, which will have a random value from 1 to 6. On success, the sent Message is returned.
+ * @method static ServerResponse sendChatAction(array $data)                  Use this method when you need to tell the user that something is happening on the bot's side. The status is set for 5 seconds or less (when a message arrives from your bot, Telegram clients clear its typing status). Returns True on success.
+ * @method static ServerResponse getUserProfilePhotos(array $data)            Use this method to get a list of profile pictures for a user. Returns a UserProfilePhotos object.
+ * @method static ServerResponse getFile(array $data)                         Use this method to get basic info about a file and prepare it for downloading. For the moment, bots can download files of up to 20MB in size. On success, a File object is returned. The file can then be downloaded via the link https://api.telegram.org/file/bot<token>/<file_path>, where <file_path> is taken from the response. It is guaranteed that the link will be valid for at least 1 hour. When the link expires, a new one can be requested by calling getFile again.
+ * @method static ServerResponse kickChatMember(array $data)                  Use this method to kick a user from a group, a supergroup or a channel. In the case of supergroups and channels, the user will not be able to return to the group on their own using invite links, etc., unless unbanned first. The bot must be an administrator in the chat for this to work and must have the appropriate admin rights. Returns True on success.
+ * @method static ServerResponse unbanChatMember(array $data)                 Use this method to unban a previously kicked user in a supergroup or channel. The user will not return to the group or channel automatically, but will be able to join via link, etc. The bot must be an administrator for this to work. Returns True on success.
+ * @method static ServerResponse restrictChatMember(array $data)              Use this method to restrict a user in a supergroup. The bot must be an administrator in the supergroup for this to work and must have the appropriate admin rights. Pass True for all permissions to lift restrictions from a user. Returns True on success.
+ * @method static ServerResponse promoteChatMember(array $data)               Use this method to promote or demote a user in a supergroup or a channel. The bot must be an administrator in the chat for this to work and must have the appropriate admin rights. Pass False for all boolean parameters to demote a user. Returns True on success.
+ * @method static ServerResponse setChatAdministratorCustomTitle(array $data) Use this method to set a custom title for an administrator in a supergroup promoted by the bot. Returns True on success.
+ * @method static ServerResponse setChatPermissions(array $data)              Use this method to set default chat permissions for all members. The bot must be an administrator in the group or a supergroup for this to work and must have the can_restrict_members admin rights. Returns True on success.
+ * @method static ServerResponse exportChatInviteLink(array $data)            Use this method to generate a new invite link for a chat. Any previously generated link is revoked. The bot must be an administrator in the chat for this to work and must have the appropriate admin rights. Returns the new invite link as String on success.
+ * @method static ServerResponse setChatPhoto(array $data)                    Use this method to set a new profile photo for the chat. Photos can't be changed for private chats. The bot must be an administrator in the chat for this to work and must have the appropriate admin rights. Returns True on success.
+ * @method static ServerResponse deleteChatPhoto(array $data)                 Use this method to delete a chat photo. Photos can't be changed for private chats. The bot must be an administrator in the chat for this to work and must have the appropriate admin rights. Returns True on success.
+ * @method static ServerResponse setChatTitle(array $data)                    Use this method to change the title of a chat. Titles can't be changed for private chats. The bot must be an administrator in the chat for this to work and must have the appropriate admin rights. Returns True on success.
+ * @method static ServerResponse setChatDescription(array $data)              Use this method to change the description of a group, a supergroup or a channel. The bot must be an administrator in the chat for this to work and must have the appropriate admin rights. Returns True on success.
+ * @method static ServerResponse pinChatMessage(array $data)                  Use this method to pin a message in a supergroup or a channel. The bot must be an administrator in the chat for this to work and must have the ‘can_pin_messages’ admin right in the supergroup or ‘can_edit_messages’ admin right in the channel. Returns True on success.
+ * @method static ServerResponse unpinChatMessage(array $data)                Use this method to unpin a message in a supergroup or a channel. The bot must be an administrator in the chat for this to work and must have the ‘can_pin_messages’ admin right in the supergroup or ‘can_edit_messages’ admin right in the channel. Returns True on success.
+ * @method static ServerResponse leaveChat(array $data)                       Use this method for your bot to leave a group, supergroup or channel. Returns True on success.
+ * @method static ServerResponse getChat(array $data)                         Use this method to get up to date information about the chat (current name of the user for one-on-one conversations, current username of a user, group or channel, etc.). Returns a Chat object on success.
+ * @method static ServerResponse getChatAdministrators(array $data)           Use this method to get a list of administrators in a chat. On success, returns an Array of ChatMember objects that contains information about all chat administrators except other bots. If the chat is a group or a supergroup and no administrators were appointed, only the creator will be returned.
  * @method static ServerResponse getChatMembersCount(array $data)             Use this method to get the number of members in a chat. Returns Int on success.
- * @method static ServerResponse getChatMember(array $data)                   Use this method to get information about a member of a chat. Returns a ChatMember
- *     object on success.
- * @method static ServerResponse setChatStickerSet(array $data)               Use this method to set a new group sticker set for a supergroup. The bot must be
- *     an administrator in the chat for this to work and must have the appropriate admin rights. Use the field can_set_sticker_set optionally returned in
- *     getChat requests to check if the bot can use this method. Returns True on success.
- * @method static ServerResponse deleteChatStickerSet(array $data)            Use this method to delete a group sticker set from a supergroup. The bot must be
- *     an administrator in the chat for this to work and must have the appropriate admin rights. Use the field can_set_sticker_set optionally returned in
- *     getChat requests to check if the bot can use this method. Returns True on success.
- * @method static ServerResponse answerCallbackQuery(array $data)             Use this method to send answers to callback queries sent from inline keyboards.
- *     The answer will be displayed to the user as a notification at the top of the chat screen or as an alert. On success, True is returned.
+ * @method static ServerResponse getChatMember(array $data)                   Use this method to get information about a member of a chat. Returns a ChatMember object on success.
+ * @method static ServerResponse setChatStickerSet(array $data)               Use this method to set a new group sticker set for a supergroup. The bot must be an administrator in the chat for this to work and must have the appropriate admin rights. Use the field can_set_sticker_set optionally returned in getChat requests to check if the bot can use this method. Returns True on success.
+ * @method static ServerResponse deleteChatStickerSet(array $data)            Use this method to delete a group sticker set from a supergroup. The bot must be an administrator in the chat for this to work and must have the appropriate admin rights. Use the field can_set_sticker_set optionally returned in getChat requests to check if the bot can use this method. Returns True on success.
+ * @method static ServerResponse answerCallbackQuery(array $data)             Use this method to send answers to callback queries sent from inline keyboards. The answer will be displayed to the user as a notification at the top of the chat screen or as an alert. On success, True is returned.
  * @method static ServerResponse answerInlineQuery(array $data)               Use this method to send answers to an inline query. On success, True is returned.
- * @method static ServerResponse setMyCommands(array $data)                   Use this method to change the list of the bot's commands. Returns True on
- *     success.
- * @method static ServerResponse getMyCommands()                              Use this method to get the current list of the bot's commands. Requires no
- *     parameters. Returns Array of BotCommand on success.
- * @method static ServerResponse editMessageText(array $data)                 Use this method to edit text and game messages sent by the bot or via the bot
- *     (for inline bots). On success, if edited message is sent by the bot, the edited Message is returned, otherwise True is returned.
- * @method static ServerResponse editMessageCaption(array $data)              Use this method to edit captions of messages sent by the bot or via the bot (for
- *     inline bots). On success, if edited message is sent by the bot, the edited Message is returned, otherwise True is returned.
- * @method static ServerResponse editMessageMedia(array $data)                Use this method to edit audio, document, photo, or video messages. On success, if
- *     the edited message was sent by the bot, the edited Message is returned, otherwise True is returned.
- * @method static ServerResponse editMessageReplyMarkup(array $data)          Use this method to edit only the reply markup of messages sent by the bot or via
- *     the bot (for inline bots). On success, if edited message is sent by the bot, the edited Message is returned, otherwise True is returned.
- * @method static ServerResponse stopPoll(array $data)                        Use this method to stop a poll which was sent by the bot. On success, the stopped
- *     Poll with the final results is returned.
- * @method static ServerResponse deleteMessage(array $data)                   Use this method to delete a message, including service messages, with certain
- *     limitations. Returns True on success.
- * @method static ServerResponse getStickerSet(array $data)                   Use this method to get a sticker set. On success, a StickerSet object is
- *     returned.
- * @method static ServerResponse uploadStickerFile(array $data)               Use this method to upload a .png file with a sticker for later use in
- *     createNewStickerSet and addStickerToSet methods (can be used multiple times). Returns the uploaded File on success.
- * @method static ServerResponse createNewStickerSet(array $data)             Use this method to create new sticker set owned by a user. The bot will be able
- *     to edit the created sticker set. Returns True on success.
- * @method static ServerResponse addStickerToSet(array $data)                 Use this method to add a new sticker to a set created by the bot. Returns True on
- *     success.
- * @method static ServerResponse setStickerPositionInSet(array $data)         Use this method to move a sticker in a set created by the bot to a specific
- *     position. Returns True on success.
- * @method static ServerResponse deleteStickerFromSet(array $data)            Use this method to delete a sticker from a set created by the bot. Returns True
- *     on success.
- * @method static ServerResponse setStickerSetThumb(array $data)              Use this method to set the thumbnail of a sticker set. Animated thumbnails can be
- *     set for animated sticker sets only. Returns True on success.
+ * @method static ServerResponse setMyCommands(array $data)                   Use this method to change the list of the bot's commands. Returns True on success.
+ * @method static ServerResponse getMyCommands()                              Use this method to get the current list of the bot's commands. Requires no parameters. Returns Array of BotCommand on success.
+ * @method static ServerResponse editMessageText(array $data)                 Use this method to edit text and game messages sent by the bot or via the bot (for inline bots). On success, if edited message is sent by the bot, the edited Message is returned, otherwise True is returned.
+ * @method static ServerResponse editMessageCaption(array $data)              Use this method to edit captions of messages sent by the bot or via the bot (for inline bots). On success, if edited message is sent by the bot, the edited Message is returned, otherwise True is returned.
+ * @method static ServerResponse editMessageMedia(array $data)                Use this method to edit audio, document, photo, or video messages. On success, if the edited message was sent by the bot, the edited Message is returned, otherwise True is returned.
+ * @method static ServerResponse editMessageReplyMarkup(array $data)          Use this method to edit only the reply markup of messages sent by the bot or via the bot (for inline bots). On success, if edited message is sent by the bot, the edited Message is returned, otherwise True is returned.
+ * @method static ServerResponse stopPoll(array $data)                        Use this method to stop a poll which was sent by the bot. On success, the stopped Poll with the final results is returned.
+ * @method static ServerResponse deleteMessage(array $data)                   Use this method to delete a message, including service messages, with certain limitations. Returns True on success.
+ * @method static ServerResponse getStickerSet(array $data)                   Use this method to get a sticker set. On success, a StickerSet object is returned.
+ * @method static ServerResponse uploadStickerFile(array $data)               Use this method to upload a .png file with a sticker for later use in createNewStickerSet and addStickerToSet methods (can be used multiple times). Returns the uploaded File on success.
+ * @method static ServerResponse createNewStickerSet(array $data)             Use this method to create new sticker set owned by a user. The bot will be able to edit the created sticker set. Returns True on success.
+ * @method static ServerResponse addStickerToSet(array $data)                 Use this method to add a new sticker to a set created by the bot. Returns True on success.
+ * @method static ServerResponse setStickerPositionInSet(array $data)         Use this method to move a sticker in a set created by the bot to a specific position. Returns True on success.
+ * @method static ServerResponse deleteStickerFromSet(array $data)            Use this method to delete a sticker from a set created by the bot. Returns True on success.
+ * @method static ServerResponse setStickerSetThumb(array $data)              Use this method to set the thumbnail of a sticker set. Animated thumbnails can be set for animated sticker sets only. Returns True on success.
  * @method static ServerResponse sendInvoice(array $data)                     Use this method to send invoices. On success, the sent Message is returned.
- * @method static ServerResponse answerShippingQuery(array $data)             If you sent an invoice requesting a shipping address and the parameter
- *     is_flexible was specified, the Bot API will send an Update with a shipping_query field to the bot. Use this method to reply to shipping queries. On
- *     success, True is returned.
- * @method static ServerResponse answerPreCheckoutQuery(array $data)          Once the user has confirmed their payment and shipping details, the Bot API sends
- *     the final confirmation in the form of an Update with the field pre_checkout_query. Use this method to respond to such pre-checkout queries. On success,
- *     True is returned.
- * @method static ServerResponse setPassportDataErrors(array $data)           Informs a user that some of the Telegram Passport elements they provided contains
- *     errors. The user will not be able to re-submit their Passport to you until the errors are fixed (the contents of the field for which you returned the
- *     error must change). Returns True on success. Use this if the data submitted by the user doesn't satisfy the standards your service requires for any
- *     reason. For example, if a birthday date seems invalid, a submitted document is blurry, a scan shows evidence of tampering, etc. Supply some details in
- *     the error message to make sure the user knows how to correct the issues.
+ * @method static ServerResponse answerShippingQuery(array $data)             If you sent an invoice requesting a shipping address and the parameter is_flexible was specified, the Bot API will send an Update with a shipping_query field to the bot. Use this method to reply to shipping queries. On success, True is returned.
+ * @method static ServerResponse answerPreCheckoutQuery(array $data)          Once the user has confirmed their payment and shipping details, the Bot API sends the final confirmation in the form of an Update with the field pre_checkout_query. Use this method to respond to such pre-checkout queries. On success, True is returned.
+ * @method static ServerResponse setPassportDataErrors(array $data)           Informs a user that some of the Telegram Passport elements they provided contains errors. The user will not be able to re-submit their Passport to you until the errors are fixed (the contents of the field for which you returned the error must change). Returns True on success. Use this if the data submitted by the user doesn't satisfy the standards your service requires for any reason. For example, if a birthday date seems invalid, a submitted document is blurry, a scan shows evidence of tampering, etc. Supply some details in the error message to make sure the user knows how to correct the issues.
  * @method static ServerResponse sendGame(array $data)                        Use this method to send a game. On success, the sent Message is returned.
- * @method static ServerResponse setGameScore(array $data)                    Use this method to set the score of the specified user in a game. On success, if
- *     the message was sent by the bot, returns the edited Message, otherwise returns True. Returns an error, if the new score is not greater than the user's
- *     current score in the chat and force is False.
- * @method static ServerResponse getGameHighScores(array $data)               Use this method to get data for high score tables. Will return the score of the
- *     specified user and several of his neighbors in a game. On success, returns an Array of GameHighScore objects.
+ * @method static ServerResponse setGameScore(array $data)                    Use this method to set the score of the specified user in a game. On success, if the message was sent by the bot, returns the edited Message, otherwise returns True. Returns an error, if the new score is not greater than the user's current score in the chat and force is False.
+ * @method static ServerResponse getGameHighScores(array $data)               Use this method to get data for high score tables. Will return the score of the specified user and several of his neighbors in a game. On success, returns an Array of GameHighScore objects.
  */
 class Request
 {
@@ -221,11 +134,11 @@ class Request
     private static $limiter_interval;
 
     /**
-     * Get the current action that is being executed
+     * The current action that is being executed
      *
      * @var string
      */
-    private static $current_action;
+    private static $current_action = '';
 
     /**
      * Available actions to send
@@ -353,7 +266,7 @@ class Request
      *
      * @param Telegram $telegram
      */
-    public static function initialize(Telegram $telegram)
+    public static function initialize(Telegram $telegram): void
     {
         self::$telegram = $telegram;
         self::setClient(self::$client ?: new Client(['base_uri' => self::$api_base_uri]));
@@ -364,7 +277,7 @@ class Request
      *
      * @param ClientInterface $client
      */
-    public static function setClient(ClientInterface $client)
+    public static function setClient(ClientInterface $client): void
     {
         self::$client = $client;
     }
@@ -373,19 +286,13 @@ class Request
      * Set input from custom input or stdin and return it
      *
      * @return string
-     * @throws TelegramException
      */
-    public static function getInput()
+    public static function getInput(): string
     {
         // First check if a custom input has been set, else get the PHP input.
         $input = self::$telegram->getCustomInput();
         if (empty($input)) {
             $input = file_get_contents('php://input');
-        }
-
-        // Make sure we have a string to work with.
-        if (!is_string($input)) {
-            throw new TelegramException('Input must be a string!');
         }
 
         TelegramLog::update($input);
@@ -400,7 +307,7 @@ class Request
      *
      * @return array Fake response data
      */
-    public static function generateGeneralFakeServerResponse(array $data = [])
+    public static function generateGeneralFakeServerResponse(array $data = []): array
     {
         //PARAM BINDED IN PHPUNIT TEST FOR TestServerResponse.php
         //Maybe this is not the best possible implementation
@@ -414,7 +321,7 @@ class Request
             $fake_response['result'] = true;
         }
 
-        //some data to let iniatilize the class method SendMessage
+        //some data to initialize the class method SendMessage
         if (isset($data['chat_id'])) {
             $data['message_id'] = '1234';
             $data['date']       = '1441378360';
@@ -442,7 +349,7 @@ class Request
      * @return array
      * @throws TelegramException
      */
-    private static function setUpRequestParams(array $data)
+    private static function setUpRequestParams(array $data): array
     {
         $has_resource = false;
         $multipart    = [];
@@ -509,7 +416,7 @@ class Request
      * @return mixed
      * @throws TelegramException
      */
-    private static function mediaInputHelper($item, &$has_resource, array &$multipart)
+    private static function mediaInputHelper($item, bool &$has_resource, array &$multipart)
     {
         $was_array = is_array($item);
         $was_array || $item = [$item];
@@ -528,7 +435,7 @@ class Request
 
             foreach ($possible_medias as $type => $media) {
                 // Allow absolute paths to local files.
-                if (is_string($media) && file_exists($media)) {
+                if (is_string($media) && strpos($media, 'attach://') !== 0 && file_exists($media)) {
                     $media = new Stream(self::encodeFile($media));
                 }
 
@@ -554,7 +461,7 @@ class Request
      *
      * @return string
      */
-    public static function getCurrentAction()
+    public static function getCurrentAction(): string
     {
         return self::$current_action;
     }
@@ -568,10 +475,8 @@ class Request
      * @return string Result of the HTTP Request
      * @throws TelegramException
      */
-    public static function execute($action, array $data = [])
+    public static function execute(string $action, array $data = []): string
     {
-        $result                  = null;
-        $response                = null;
         $request_params          = self::setUpRequestParams($data);
         $request_params['debug'] = TelegramLog::getDebugLogTempStream();
 
@@ -589,13 +494,13 @@ class Request
         } catch (RequestException $e) {
             $response = null;
             $result   = $e->getResponse() ? (string) $e->getResponse()->getBody() : '';
-        } finally {
-            //Logging verbose debug output
-            if (TelegramLog::$always_log_request_and_response || $response === null) {
-                TelegramLog::debug('Request data:' . PHP_EOL . print_r($data, true));
-                TelegramLog::debug('Response data:' . PHP_EOL . $result);
-                TelegramLog::endDebugLogTempStream('Verbose HTTP Request output:' . PHP_EOL . '%s' . PHP_EOL);
-            }
+        }
+
+        //Logging verbose debug output
+        if (TelegramLog::$always_log_request_and_response || $response === null) {
+            TelegramLog::debug('Request data:' . PHP_EOL . print_r($data, true));
+            TelegramLog::debug('Response data:' . PHP_EOL . $result);
+            TelegramLog::endDebugLogTempStream('Verbose HTTP Request output:' . PHP_EOL . '%s' . PHP_EOL);
         }
 
         return $result;
@@ -609,7 +514,7 @@ class Request
      * @return bool
      * @throws TelegramException
      */
-    public static function downloadFile(File $file)
+    public static function downloadFile(File $file): bool
     {
         if (empty($download_path = self::$telegram->getDownloadPath())) {
             throw new TelegramException('Download path not set!');
@@ -634,7 +539,7 @@ class Request
             );
 
             return filesize($file_path) > 0;
-        } catch (RequestException $e) {
+        } catch (Throwable $e) {
             return false;
         } finally {
             //Logging verbose debug output
@@ -650,7 +555,7 @@ class Request
      * @return resource
      * @throws TelegramException
      */
-    public static function encodeFile($file)
+    public static function encodeFile(string $file)
     {
         $fp = fopen($file, 'rb');
         if ($fp === false) {
@@ -672,7 +577,7 @@ class Request
      * @return ServerResponse
      * @throws TelegramException
      */
-    public static function send($action, array $data = [])
+    public static function send(string $action, array $data = []): ServerResponse
     {
         self::ensureValidAction($action);
         self::addDummyParamIfNecessary($action, $data);
@@ -713,7 +618,7 @@ class Request
         }
 
         // Reset current action after completion.
-        self::$current_action = null;
+        self::$current_action = '';
 
         return $response;
     }
@@ -731,7 +636,7 @@ class Request
      * @param string $action
      * @param array  $data
      */
-    protected static function addDummyParamIfNecessary($action, array &$data)
+    protected static function addDummyParamIfNecessary(string $action, array &$data): void
     {
         if (in_array($action, self::$actions_need_dummy_param, true)) {
             // Can be anything, using a single letter to minimise request size.
@@ -746,7 +651,7 @@ class Request
      *
      * @throws TelegramException
      */
-    private static function ensureNonEmptyData(array $data)
+    private static function ensureNonEmptyData(array $data): void
     {
         if (count($data) === 0) {
             throw new TelegramException('Data is empty!');
@@ -760,7 +665,7 @@ class Request
      *
      * @throws TelegramException
      */
-    private static function ensureValidAction($action)
+    private static function ensureValidAction(string $action): void
     {
         if (!in_array($action, self::$actions, true)) {
             throw new TelegramException('The action "' . $action . '" doesn\'t exist!');
@@ -777,7 +682,7 @@ class Request
      * @return ServerResponse
      * @throws TelegramException
      */
-    public static function sendMessage(array $data)
+    public static function sendMessage(array $data): ServerResponse
     {
         $text = $data['text'];
 
@@ -800,27 +705,25 @@ class Request
      * @param array  $data
      *
      * @return ServerResponse
+     * @throws TelegramException
      */
-    public static function __callStatic($action, array $data)
+    public static function __callStatic(string $action, array $data)
     {
-        // Make sure to add the action being called as the first parameter to be passed.
-        array_unshift($data, $action);
-
-        // @todo Use splat operator for unpacking when we move to PHP 5.6+
-        return call_user_func_array('static::send', $data);
+        // Only argument should be the data array, ignore any others.
+        return static::send($action, reset($data));
     }
 
     /**
      * Return an empty Server Response
      *
-     * No request to telegram are sent, this function is used in commands that
-     * don't need to fire a message after execution
+     * No request is sent to Telegram.
+     * This function is used in commands that don't need to fire a message after execution
      *
      * @return ServerResponse
      */
-    public static function emptyResponse()
+    public static function emptyResponse(): ServerResponse
     {
-        return new ServerResponse(['ok' => true, 'result' => true], null);
+        return new ServerResponse(['ok' => true, 'result' => true]);
     }
 
     /**
@@ -834,10 +737,10 @@ class Request
      * @throws TelegramException
      */
     public static function sendToActiveChats(
-        $callback_function,
+        string $callback_function,
         array $data,
         array $select_chats_params
-    ) {
+    ): array {
         self::ensureValidAction($callback_function);
 
         $chats = DB::selectChats($select_chats_params);
@@ -856,12 +759,12 @@ class Request
     /**
      * Enable request limiter
      *
-     * @param bool $enable
-     * @param array   $options
+     * @param bool  $enable
+     * @param array $options
      *
      * @throws TelegramException
      */
-    public static function setLimiter($enable = true, array $options = [])
+    public static function setLimiter(bool $enable = true, array $options = []): void
     {
         if (DB::isDbConnected()) {
             $options_default = [
@@ -890,7 +793,7 @@ class Request
      *
      * @throws TelegramException
      */
-    private static function limitTelegramRequests($action, array $data = [])
+    private static function limitTelegramRequests(string $action, array $data = []): void
     {
         if (self::$limiter_enabled) {
             $limited_methods = [
@@ -928,8 +831,8 @@ class Request
                 'setPassportDataErrors',
             ];
 
-            $chat_id           = isset($data['chat_id']) ? $data['chat_id'] : null;
-            $inline_message_id = isset($data['inline_message_id']) ? $data['inline_message_id'] : null;
+            $chat_id           = $data['chat_id'] ?? null;
+            $inline_message_id = $data['inline_message_id'] ?? null;
 
             if (($chat_id || $inline_message_id) && in_array($action, $limited_methods, true)) {
                 $timeout = 60;
