@@ -14,6 +14,7 @@ namespace Longman\TelegramBot\Commands\AdminCommands;
 use Exception;
 use Longman\TelegramBot\Commands\AdminCommand;
 use Longman\TelegramBot\DB;
+use Longman\TelegramBot\Entities\ServerResponse;
 use Longman\TelegramBot\Exception\TelegramException;
 use Longman\TelegramBot\Request;
 
@@ -48,7 +49,7 @@ class DebugCommand extends AdminCommand
      * @return mixed
      * @throws TelegramException
      */
-    public function execute()
+    public function execute(): ServerResponse
     {
         $pdo     = DB::getPdo();
         $message = $this->getMessage();
